1 service per app resolves #611
