+++ pkg/util/ds/slice.go
@@ -1,11 +1,21 @@
 package ds
 
+import (
+	"golang.org/x/exp/constraints"
+)
+
 func Map[T, U any](s []T, mapper func(item T) U) []U {
 	ret := make([]U, len(s))
 	for i := range s {
 		ret[i] = mapper(s[i])
 	}
 	return ret
+}
+
+func LessFunc[E any, K constraints.Ordered](key func(e E) K) func(e1, e2 E) bool {
+	return func(e1, e2 E) bool {
+		return key(e1) < key(e2)
+	}
 }
 
 func SliceOfPtr[T any](s []T) []*T {
